+++ packages/bundle-utils/src/json.ts
@@ -1,5 +1,5 @@
 /**
- * Code generator for i18n json/json5 resource
+ * Code / AST generator for i18n json/json5 resource
  */
 
 import {
@@ -11,24 +11,30 @@ import { isString, friendlyJSONstringify } from '@intlify/shared'
 import {
   createCodeGenerator,
   generateMessageFunction,
-  mapLinesColumns
+  generateResourceAst,
+  mapLinesColumns,
+  excludeLocales
 } from './codegen'
 import { generateLegacyCode } from './legacy'
-import { RawSourceMap } from 'source-map'
 import MagicString from 'magic-string'
 
+import type { RawSourceMap } from 'source-map-js'
 import type { JSONProgram, JSONNode } from 'jsonc-eslint-parser/lib/parser/ast'
-import type { CodeGenOptions, CodeGenerator, CodeGenResult } from './codegen'
+import type {
+  CodeGenOptions,
+  CodeGenerator,
+  CodeGenResult,
+  CodeGenFunction
+} from './codegen'
 
-/**
- * @internal
- */
 export function generate(
   targetSource: string | Buffer,
   {
     type = 'plain',
     legacy = false,
+    vueVersion = 'v2.6',
     bridge = false,
+    onlyLocales = [],
     exportESM = false,
     filename = 'vue-i18n-loader.json',
     inSourceMap = undefined,
@@ -40,17 +46,17 @@ export function generate(
     onError = undefined,
     strictMessage = true,
     escapeHtml = false,
-    useClassComponent = false
+    useClassComponent = false,
+    jit = false
   }: CodeGenOptions,
   injector?: () => string
 ): CodeGenResult<JSONProgram> {
-  const target = Buffer.isBuffer(targetSource)
+  let value = Buffer.isBuffer(targetSource)
     ? targetSource.toString()
     : targetSource
   // const value = JSON.stringify(JSON.parse(target))
   //   .replace(/\u2028/g, '\\u2028') // line separator
   //   .replace(/\u2029/g, '\\u2029') // paragraph separator
-  const value = target
 
   const options = {
     type,
@@ -67,16 +73,29 @@ export function generate(
     onError,
     strictMessage,
     escapeHtml,
-    useClassComponent
+    useClassComponent,
+    jit
   } as CodeGenOptions
   const generator = createCodeGenerator(options)
 
-  const ast = parseJSON(value, { filePath: filename })
+  let ast = parseJSON(value, { filePath: filename })
+
+  if (!locale && type === 'sfc' && onlyLocales?.length) {
+    const messages = getStaticJSONValue(ast) as Record<string, unknown>
+
+    value = JSON.stringify(
+      excludeLocales({
+        messages,
+        onlyLocales
+      })
+    )
+    ast = parseJSON(value, { filePath: filename })
+  }
 
   // for vue 2.x
   if (legacy && type === 'sfc') {
     const gen = () => friendlyJSONstringify(getStaticJSONValue(ast))
-    const code = generateLegacyCode(options, gen)
+    const code = generateLegacyCode({ isGlobal, vueVersion }, gen)
     const s = new MagicString(code)
     return {
       ast,
@@ -89,7 +108,7 @@ export function generate(
     }
   }
 
-  const codeMaps = generateNode(generator, ast, options, injector)
+  const codeMaps = _generate(generator, ast, options, injector)
 
   const { code, map } = generator.context()
   // if (map) {
@@ -99,9 +118,10 @@ export function generate(
   //   })
   // }
   // prettier-ignore
-  const newMap = map
-    ? mapLinesColumns((map as any).toJSON(), codeMaps, inSourceMap) || null // eslint-disable-line @typescript-eslint/no-explicit-any
-    : null
+  const newMap =
+    map && !jit
+      ? mapLinesColumns((map as any).toJSON(), codeMaps, inSourceMap) || null // eslint-disable-line @typescript-eslint/no-explicit-any
+      : null
   return {
     ast,
     code,
@@ -109,10 +129,10 @@ export function generate(
   }
 }
 
-function generateNode(
+function _generate(
   generator: CodeGenerator,
   node: JSONProgram,
-  options: CodeGenOptions,
+  options: CodeGenOptions = {},
   injector?: () => string
 ): Map<string, RawSourceMap> {
   const propsCountStack = [] as number[]
@@ -127,8 +147,13 @@ function generateNode(
     sourceMap,
     isGlobal,
     locale,
-    useClassComponent
+    useClassComponent,
+    jit
   } = options
+
+  const codegenFn: CodeGenFunction = jit
+    ? generateResourceAst
+    : generateMessageFunction
 
   const componentNamespace = '_Component'
 
@@ -137,7 +162,7 @@ function generateNode(
       switch (node.type) {
         case 'Program':
           if (type === 'plain') {
-            generator.push(`export default `)
+            generator.push(`const resource = `)
           } else if (type === 'sfc') {
             // for 'sfc'
             const variableName =
@@ -155,7 +180,7 @@ function generateNode(
             const componentVariable = bridge
               ? `Component.options || Component`
               : useClassComponent
-                ? `Component.__o || Component`
+                ? `Component.__o || Component.__vccOpts || Component`
                 : `Component`
             // prettier-ignore
             generator.pushline(`const ${componentNamespace} = ${componentVariable}`)
@@ -192,11 +217,7 @@ function generateNode(
             if (isString(value)) {
               generator.push(`${JSON.stringify(name)}: `)
               pathStack.push(name.toString())
-              const { code, map } = generateMessageFunction(
-                value,
-                options,
-                pathStack
-              )
+              const { code, map } = codegenFn(value, options, pathStack)
               sourceMap && map != null && codeMaps.set(value, map)
               generator.push(`${code}`, node.value, value)
             } else {
@@ -204,11 +225,7 @@ function generateNode(
                 const strValue = JSON.stringify(value)
                 generator.push(`${JSON.stringify(name)}: `)
                 pathStack.push(name.toString())
-                const { code, map } = generateMessageFunction(
-                  strValue,
-                  options,
-                  pathStack
-                )
+                const { code, map } = codegenFn(strValue, options, pathStack)
                 sourceMap && map != null && codeMaps.set(strValue, map)
                 generator.push(`${code}`, node.value, strValue)
               } else {
@@ -253,21 +270,13 @@ function generateNode(
             if (node.type === 'JSONLiteral') {
               const value = node.value
               if (isString(value)) {
-                const { code, map } = generateMessageFunction(
-                  value,
-                  options,
-                  pathStack
-                )
+                const { code, map } = codegenFn(value, options, pathStack)
                 sourceMap && map != null && codeMaps.set(value, map)
                 generator.push(`${code}`, node, value)
               } else {
                 if (forceStringify) {
                   const strValue = JSON.stringify(value)
-                  const { code, map } = generateMessageFunction(
-                    strValue,
-                    options,
-                    pathStack
-                  )
+                  const { code, map } = codegenFn(strValue, options, pathStack)
                   sourceMap && map != null && codeMaps.set(strValue, map)
                   generator.push(`${code}`, node, strValue)
                 } else {
@@ -300,6 +309,9 @@ function generateNode(
             }
             generator.deindent()
             generator.pushline(`}`)
+          } else if (type === 'plain') {
+            generator.push(`\n`)
+            generator.push('export default resource')
           }
           break
         case 'JSONObjectExpression':