+++ packages/bundle-utils/test/generator/codegen.test.ts
@@ -1,4 +1,29 @@
-import { generateMessageFunction } from '../../src/codegen'
+import { generateMessageFunction, generateResourceAst } from '../../src/codegen'
+
+describe('generateResourceAst', () => {
+  test('development', () => {
+    expect(generateResourceAst('hello')).toMatchSnapshot()
+  })
+  test('production', () => {
+    expect(
+      generateResourceAst('hello', {
+        env: 'production',
+        jit: true,
+        minify: true
+      })
+    ).toMatchSnapshot()
+  })
+  test('syntax error', () => {
+    const errors: string[] = []
+    const { code } = generateMessageFunction(`|`, {
+      onError(msg) {
+        errors.push(msg)
+      }
+    })
+    expect(errors.length).toBe(1)
+    expect(code).toBe('`|`')
+  })
+})
 
 describe('generateMessageFunction', () => {
   test('development', () => {
@@ -23,7 +48,7 @@ describe('generateMessageFunction', () => {
       }
     })
     expect(errors.length).toBe(1)
-    expect(code).toBe(`|`)
+    expect(code).toBe('`|`')
   })
 
   describe('strictMessage', () => {