+++ packages/bundle-utils/test/generator/js.test.ts
@@ -135,7 +135,7 @@ test('useClassComponent', async () => {
 
   expect(validateSyntax(code)).toBe(true)
   expect(code).toMatchSnapshot('code')
-  expect(code).toContain('Component.__o || Component')
+  expect(code).toContain('Component.__o || Component.__vccOpts || Component')
   expect(code).toContain('_Component.__i18n')
   expect(map).toMatchSnapshot('map')
 })
@@ -271,4 +271,15 @@ test('include function', async () => {
 
   expect(code).toMatchSnapshot('code')
   expect(map).toMatchSnapshot('map')
+})
+
+test('AST code generation', async () => {
+  const { source } = await readFile('./fixtures/codegen/simple.js')
+  const { code } = generate(source, {
+    jit: true,
+    env: 'production'
+  })
+
+  expect(validateSyntax(code)).toBe(true)
+  expect(code).toMatchSnapshot('code')
 })