+++ packages/bundle-utils/test/generator/yaml.test.ts
@@ -115,7 +115,7 @@ test('useClassComponent', async () => {
 
   expect(validateSyntax(code)).toBe(true)
   expect(code).toMatchSnapshot('code')
-  expect(code).toContain('Component.__o || Component')
+  expect(code).toContain('Component.__o || Component.__vccOpts || Component')
   expect(code).toContain('_Component.__i18n')
   expect(map).toMatchSnapshot('map')
 })
@@ -176,5 +176,15 @@ test('html tag in message', async () => {
   })
 
   expect(errors).toMatchSnapshot('errors')
+  expect(code).toMatchSnapshot('code')
+})
+
+test('AST code generation', async () => {
+  const { source } = await readFile(`./fixtures/codegen/complex.yaml`)
+  const { code } = generate(source, {
+    jit: true,
+    env: 'production'
+  })
+
   expect(code).toMatchSnapshot('code')
 })