+++ packages/unplugin-vue-i18n/test/vite/custom-block.test.ts
@@ -1,5 +1,5 @@
 import { bundleVite, bundleAndRun } from '../utils'
-import { createMessageContext } from '@intlify/core-base'
+import { createMessageContext, isMessageAST } from '@intlify/core-base'
 
 test('basic', async () => {
   const { module } = await bundleAndRun('basic.vue', bundleVite)
@@ -7,6 +7,33 @@ test('basic', async () => {
   const i18n = module.__i18n.pop()
   expect(i18n.locale).toEqual('')
   expect(i18n.resource.en.hello(createMessageContext())).toEqual('hello world!')
+})
+
+test('json: exclude locales', async () => {
+  const { module } = await bundleAndRun('basic.vue', bundleVite, {
+    onlyLocales: ['ja']
+  })
+  expect(module.__i18n).toMatchSnapshot()
+  const i18n = module.__i18n.pop()
+  expect(i18n.resource.en).toBeUndefined()
+})
+
+test('yaml: exclude locales', async () => {
+  const { module } = await bundleAndRun('yaml.vue', bundleVite, {
+    onlyLocales: ['ja']
+  })
+  expect(module.__i18n).toMatchSnapshot()
+  const i18n = module.__i18n.pop()
+  expect(i18n.resource.en).toBeUndefined()
+})
+
+test('json5: exclude locales', async () => {
+  const { module } = await bundleAndRun('json5.vue', bundleVite, {
+    onlyLocales: ['ja']
+  })
+  expect(module.__i18n).toMatchSnapshot()
+  const i18n = module.__i18n.pop()
+  expect(i18n.resource.en).toBeUndefined()
 })
 
 test('yaml', async () => {
@@ -137,6 +164,23 @@ test('global scope and import', async () => {
   expect(g.resource.en.hello(createMessageContext())).toEqual('hello world!')
 })
 
+test('legacy for Vue 2.6 and before', async () => {
+  const { code } = await bundleAndRun('basic.vue', bundleVite, {
+    legacy: true
+  })
+  expect(code).toMatchSnapshot()
+  expect(code).toMatch('Component.options.__i18n =')
+})
+
+test('legacy for Vue 2.7', async () => {
+  const { module } = await bundleAndRun('basic.vue', bundleVite, {
+    legacy: true,
+    vueVersion: 'v2.7'
+  })
+  const i18n = JSON.parse(module.__i18n.pop())
+  expect(i18n.en.hello).toEqual('hello world!')
+})
+
 test('array', async () => {
   const { module } = await bundleAndRun('array.vue', bundleVite)
   expect(module.__i18n).toMatchSnapshot()
@@ -144,4 +188,15 @@ test('array', async () => {
   expect(i18n.locale).toEqual('')
   expect(i18n.resource.en.foo[0][0](createMessageContext())).toEqual('bar')
   expect(i18n.resource.en.foo[1](createMessageContext())).toEqual('baz')
+})
+
+test('jitCompilation', async () => {
+  const { module } = await bundleAndRun('basic.vue', bundleVite, {
+    jitCompilation: true,
+    env: 'production'
+  })
+  expect(module.__i18n).toMatchSnapshot()
+  const i18n = module.__i18n.pop()
+  expect(i18n.locale).toEqual('')
+  expect(isMessageAST(i18n.resource.en.hello)).toBe(true)
 })