+++ packages/unplugin-vue-i18n/CHANGELOG.md
@@ -1,4 +1,132 @@
 
+## @intlify/unplugin-vue-i18n@1.6.0 (2023-12-13)
+
+#### :star: Features
+* [#332](https://github.com/intlify/bundle-tools/pull/332) feat(unplugin-vue-i18n): support compatibility i18n custom blocks for legacy vue-i18n v8.x ([@kazupon](https://github.com/kazupon))
+
+#### Committers: 1
+- kazuya kawaguchi ([@kazupon](https://github.com/kazupon))
+
+
+## @intlify/unplugin-vue-i18n@1.5.0 (2023-11-02)
+
+#### :star: Features
+* [#274](https://github.com/intlify/bundle-tools/pull/274) feat: merging of locale files with the same name for `@intlify/unplugin-vue-i18n/messages` ([@DanielleHuisman](https://github.com/DanielleHuisman))
+
+#### Committers: 1
+- Daniëlle Huisman ([@DanielleHuisman](https://github.com/DanielleHuisman))
+
+
+## @intlify/unplugin-vue-i18n@1.4.0 (2023-09-29)
+
+#### :star: Features
+* [#315](https://github.com/intlify/bundle-tools/pull/315) feat: `ssr` option ([@kazupon](https://github.com/kazupon))
+
+#### Committers: 1
+- kazuya kawaguchi ([@kazupon](https://github.com/kazupon))
+
+
+## @intlify/unplugin-vue-i18n@1.3.0 (2023-09-29)
+
+#### :star: Features
+* [#301](https://github.com/intlify/bundle-tools/pull/301) feat(unplugin-vue-i18n): exclude unused locales from the bundle ([@imslepov](https://github.com/imslepov))
+
+#### Committers: 1
+- Daniel Slepov ([@imslepov](https://github.com/imslepov))
+
+
+
+
+
+
+
+
+
+
+## @intlify/unplugin-vue-i18n@0.13.0 (2023-08-30)
+
+#### :star: Features
+* [#288](https://github.com/intlify/bundle-tools/pull/288) feat(unplugin-vue-i18n): support `dropMessageCompiler` option ([@kazupon](https://github.com/kazupon))
+
+#### Committers: 1
+- kazuya kawaguchi ([@kazupon](https://github.com/kazupon))
+
+
+## @intlify/unplugin-vue-i18n@0.13.0 (2023-08-30)
+
+#### :star: Features
+* [#287](https://github.com/intlify/bundle-tools/pull/287) feat(bundle-utils): support AST minify option ([@kazupon](https://github.com/kazupon))
+
+#### :bug: Bug Fixes
+* [#286](https://github.com/intlify/bundle-tools/pull/286) fix: Use `defaultSFCLang` setting for src files instead of falling back to json ([@Lutymane](https://github.com/Lutymane))
+
+#### Committers: 2
+- [@Lutymane](https://github.com/Lutymane)
+- kazuya kawaguchi ([@kazupon](https://github.com/kazupon))
+
+
+## @intlify/unplugin-vue-i18n@0.12.3 (2023-08-17)
+
+#### :zap: Improvement Features
+* [#272](https://github.com/intlify/bundle-tools/pull/272) Set `messages` type to the vue-i18n type instead of any ([@ferferga](https://github.com/ferferga))
+
+#### Committers: 1
+- Fernando Fernández ([@ferferga](https://github.com/ferferga))
+
+
+
+
+## @intlify/unplugin-vue-i18n@0.12.1 (2023-07-07)
+
+#### :bug: Bug Fixes
+* [#279](https://github.com/intlify/bundle-tools/pull/279) fix(unplugin-vue-i18n): sourcemap warning ([@kazupon](https://github.com/kazupon))
+
+#### Committers: 1
+- kazuya kawaguchi ([@kazupon](https://github.com/kazupon))
+
+
+## @intlify/unplugin-vue-i18n@0.12.0 (2023-07-01)
+
+#### :star: Features
+* [#277](https://github.com/intlify/bundle-tools/pull/277) feat(unplugin-vue-i18n): support jit compilation ([@kazupon](https://github.com/kazupon))
+
+#### Committers: 1
+- kazuya kawaguchi ([@kazupon](https://github.com/kazupon))
+
+
+## @intlify/unplugin-vue-i18n@0.11.0 (2023-05-26)
+
+#### :zap: Improvement Features
+* [#271](https://github.com/intlify/bundle-tools/pull/271) fix(unplugin-vue-i18n): bump intlify/bundle-utils ([@kazupon](https://github.com/kazupon))
+
+#### Committers: 1
+- kazuya kawaguchi ([@kazupon](https://github.com/kazupon))
+
+
+## @intlify/unplugin-vue-i18n@0.10.1 (2023-05-26)
+
+#### :bug: Bug Fixes
+* [#270](https://github.com/intlify/bundle-tools/pull/270) fix: escape string, when there's a syntax failure ([@kazupon](https://github.com/kazupon))
+
+#### :pencil: Documentation
+* [#261](https://github.com/intlify/bundle-tools/pull/261) fix: default plugin locale path ([@akires47](https://github.com/akires47))
+
+#### Committers: 2
+- Tomas Barrera ([@akires47](https://github.com/akires47))
+- kazuya kawaguchi ([@kazupon](https://github.com/kazupon))
+
+
+## @intlify/unplugin-vue-i18n@0.10.1 (2023-03-28)
+
+#### :zap: Improvement Features
+* [#254](https://github.com/intlify/bundle-tools/pull/254) fix(bundle-utils): improve plain code generation ([@kazupon](https://github.com/kazupon))
+* [#253](https://github.com/intlify/bundle-tools/pull/253) fix: types exports for "moduleResolution": "node16" ([@RebeccaStevens](https://github.com/RebeccaStevens))
+
+#### Committers: 2
+- Rebecca Stevens ([@RebeccaStevens](https://github.com/RebeccaStevens))
+- kazuya kawaguchi ([@kazupon](https://github.com/kazupon))
+
+
 ## @intlify/unplugin-vue-i18n@0.10.0 (2023-03-22)
 
 #### :star: Features